+++ src/amg/azext_amg/_params.py
@@ -9,7 +9,7 @@ def load_arguments(self, _):
 
     from knack.arguments import CLIArgumentType
     from azure.cli.core.commands.parameters import tags_type, get_three_state_flag, get_enum_type
-    from azure.cli.core.commands.validators import get_default_location_from_resource_group
+    from azure.cli.core.commands.validators import get_default_location_from_resource_group, validate_file_or_dict
     from ._validators import process_missing_resource_group_parameter
     from azext_amg.vendored_sdks.models import ZoneRedundancy
     grafana_name_type = CLIArgumentType(options_list="--grafana-name",
@@ -29,27 +29,38 @@ def load_arguments(self, _):
         c.argument("zone_redundancy", arg_type=get_enum_type(ZoneRedundancy), help="Indicates whether or not zone redundancy should be enabled. Default: Disabled")
         c.argument("skip_system_assigned_identity", options_list=["-s", "--skip-system-assigned-identity"], arg_type=get_three_state_flag(), help="Do not enable system assigned identity")
         c.argument("skip_role_assignments", arg_type=get_three_state_flag(), help="Do not create role assignments for managed identity and the current login user")
+        c.argument("principal_ids", nargs="+", help="space-separated Azure AD object ids for users, groups, etc to be made as Grafana Admins. Once provided, CLI won't make the current logon user as Grafana Admin")
 
     with self.argument_context("grafana dashboard") as c:
         c.argument("uid", options_list=["--dashboard"], help="dashboard uid")
-        c.argument("definition", help="The complete dashboard model in json string, a path or url to a file with such content")
         c.argument("title", help="title of a dashboard")
         c.argument('overwrite', arg_type=get_three_state_flag(), help='Overwrite a dashboard with same uid')
+
+    with self.argument_context("grafana dashboard create") as c:
+        c.argument("definition", type=validate_file_or_dict, help="The complete dashboard model in json string, a path or url to a file with such content")
+
+    with self.argument_context("grafana dashboard update") as c:
+        c.argument("definition", type=validate_file_or_dict, help="The complete dashboard model in json string, a path or url to a file with such content")
 
     with self.argument_context("grafana dashboard import") as c:
         c.argument("definition", help="The complete dashboard model in json string, Grafana gallery id, a path or url to a file with such content")
 
     with self.argument_context("grafana data-source") as c:
         c.argument("data_source", help="name, id, uid which can identify a data source. CLI will search in the order of name, id, and uid, till finds a match")
-        c.argument("definition", help="json string with data source definition, or a path to a file with such content")
+        c.argument("definition", type=validate_file_or_dict, help="json string with data source definition, or a path to a file with such content")
+
+    with self.argument_context("grafana notification-channel") as c:
+        c.argument("notification_channel", help="id, uid which can identify a data source. CLI will search in the order of id, and uid, till finds a match")
+        c.argument("definition", type=validate_file_or_dict, help="json string with notification channel definition, or a path to a file with such content")
+        c.argument("short", action='store_true', help="list notification channels in short format.")
 
     with self.argument_context("grafana data-source query") as c:
         c.argument("conditions", nargs="+", help="space-separated condition in a format of `<name>=<value>`")
         c.argument("time_from", options_list=["--from"], help="start time in iso 8601, e.g. '2022-01-02T16:15:00'. Default: 1 hour early")
         c.argument("time_to", options_list=["--to"], help="end time in iso 8601, e.g. '2022-01-02T17:15:00'. Default: current time ")
-        c.argument("max_data_points", help="Maximum amount of data points that dashboard panel can render")
+        c.argument("max_data_points", type=int, help="Maximum amount of data points that dashboard panel can render. Default: 1000")
         c.argument("query_format", help="format of the resule, e.g. table, time_series")
-        c.argument("internal_ms", help="The time interval in milliseconds of time series")
+        c.argument("internal_ms", type=int, help="The time interval in milliseconds of time series. Default: 1000")
 
     with self.argument_context("grafana folder") as c:
         c.argument("title", help="title of the folder")