+++ src/amg/azext_amg/custom.py
@@ -26,13 +26,16 @@ grafana_endpoints = {}
 
 def create_grafana(cmd, resource_group_name, grafana_name,
                    location=None, skip_system_assigned_identity=False, skip_role_assignments=False,
-                   tags=None, zone_redundancy=None):
+                   tags=None, zone_redundancy=None, principal_ids=None):
     from azure.cli.core.commands.arm import resolve_role_id
+
+    if skip_role_assignments and principal_ids:
+        raise ArgumentUsageError("--skip-role-assignments | --assignee-object-ids")
 
     client = cf_amg(cmd.cli_ctx)
     resource = {
         "sku": {
-            "name": "standard"
+            "name": "Standard"
         },
         "location": location,
         "identity": None if skip_system_assigned_identity else {"type": "SystemAssigned"},
@@ -54,9 +57,13 @@ def create_grafana(cmd, resource_group_name, grafana_name,
 
     subscription_scope = '/subscriptions/' + client._config.subscription_id  # pylint: disable=protected-access
 
-    user_principal_id = _get_login_account_principal_id(cmd.cli_ctx)
+    if not principal_ids:
+        user_principal_id = _get_login_account_principal_id(cmd.cli_ctx)
+        principal_ids = [user_principal_id]
     grafana_admin_role_id = resolve_role_id(cmd.cli_ctx, "Grafana Admin", subscription_scope)
-    _create_role_assignment(cmd.cli_ctx, user_principal_id, grafana_admin_role_id, resource.id)
+
+    for p in principal_ids:
+        _create_role_assignment(cmd.cli_ctx, p, grafana_admin_role_id, resource.id)
 
     if resource.identity:
         monitoring_reader_role_id = resolve_role_id(cmd.cli_ctx, "Monitoring Reader", subscription_scope)
@@ -75,13 +82,13 @@ def _get_login_account_principal_id(cli_ctx):
     client = GraphRbacManagementClient(cred, tenant_id,
                                        base_url=cli_ctx.cloud.endpoints.active_directory_graph_resource_id)
     assignee = profile.get_current_account_user()
-    result = list(client.users.list(filter=f"userPrincipalName eq '{assignee}'"))
+    result = list(client.users.list(filter=f"userPrincipalName eq '{assignee}' or mail eq '{assignee}'"))
     if not result:
         result = list(client.service_principals.list(
             filter=f"servicePrincipalNames/any(c:c eq '{assignee}')"))
     if not result:
         raise CLIInternalError((f"Failed to retrieve principal id for '{assignee}', which is needed to create a "
-                                f"role assignment"))
+                                f"role assignment. Consider using '--principal-ids' to bypass the lookup"))
     return result[0].object_id
 
 
@@ -159,26 +166,25 @@ def list_dashboards(cmd, grafana_name, resource_group_name=None):
 
 
 def create_dashboard(cmd, grafana_name, definition, title=None, folder=None, resource_group_name=None, overwrite=None):
-    data = _try_load_dashboard_definition(cmd, resource_group_name, grafana_name, definition, for_import=False)
-    if "dashboard" in data:
-        payload = data
+    if "dashboard" in definition:
+        payload = definition
     else:
         logger.info("Adjust input by adding 'dashboard' field")
         payload = {}
-        payload["dashboard"] = data
+        payload['dashboard'] = definition
 
     if title:
         payload['dashboard']['title'] = title
 
     if folder:
         folder = _find_folder(cmd, resource_group_name, grafana_name, folder)
-        payload['folderId'] = folder["id"]
+        payload['folderId'] = folder['id']
 
-    payload["overwrite"] = overwrite or False
+    payload['overwrite'] = overwrite or False
 
-    if "id" in payload["dashboard"]:
+    if "id" in payload['dashboard']:
         logger.warning("Removing 'id' from dashboard to prevent the error of 'Not Found'")
-        del payload["dashboard"]["id"]
+        del payload['dashboard']['id']
 
     response = _send_request(cmd, resource_group_name, grafana_name, "post", "/api/dashboards/db",
                              payload)
@@ -193,31 +199,31 @@ def update_dashboard(cmd, grafana_name, definition, folder=None, resource_group_
 
 def import_dashboard(cmd, grafana_name, definition, folder=None, resource_group_name=None, overwrite=None):
     import copy
-    data = _try_load_dashboard_definition(cmd, resource_group_name, grafana_name, definition, for_import=True)
+    data = _try_load_dashboard_definition(cmd, resource_group_name, grafana_name, definition)
     if "dashboard" in data:
         payload = data
     else:
         logger.info("Adjust input by adding 'dashboard' field")
         payload = {}
-        payload["dashboard"] = data
+        payload['dashboard'] = data
 
     if folder:
         folder = _find_folder(cmd, resource_group_name, grafana_name, folder)
-        payload['folderId'] = folder["id"]
+        payload['folderId'] = folder['id']
 
-    payload["overwrite"] = overwrite or False
+    payload['overwrite'] = overwrite or False
 
-    payload["inputs"] = []
+    payload['inputs'] = []
 
     # provide parameter values for datasource
     data_sources = list_data_sources(cmd, grafana_name, resource_group_name)
-    for parameter in payload["dashboard"].get('__inputs', []):
+    for parameter in payload['dashboard'].get('__inputs', []):
         if parameter.get("type") == "datasource":
-            match = next((d for d in data_sources if d["type"] == parameter["pluginId"]), None)
+            match = next((d for d in data_sources if d['type'] == parameter['pluginId']), None)
             if match:
                 clone = copy.deepcopy(parameter)
-                clone["value"] = match["uid"]
-                payload["inputs"].append(clone)
+                clone['value'] = match['uid']
+                payload['inputs'].append(clone)
             else:
                 logger.warning("No data source was found matching the required parameter of %s", parameter['pluginId'])
 
@@ -226,17 +232,17 @@ def import_dashboard(cmd, grafana_name, definition, folder=None, resource_group_
     return json.loads(response.content)
 
 
-def _try_load_dashboard_definition(cmd, resource_group_name, grafana_name, definition, for_import):
+def _try_load_dashboard_definition(cmd, resource_group_name, grafana_name, definition):
     import re
 
-    if for_import:
-        try:  # see whether it is a gallery id
-            int(definition)
-            response = _send_request(cmd, resource_group_name, grafana_name, "get",
-                                     "/api/gnet/dashboards/" + definition)
-            return json.loads(response.content)["json"]
-        except ValueError:
-            pass
+    try:
+        int(definition)
+        response = _send_request(cmd, resource_group_name, grafana_name, "get",
+                                 "/api/gnet/dashboards/" + str(definition))
+        definition = json.loads(response.content)["json"]
+        return definition
+    except ValueError:
+        pass
 
     if re.match(r"^[a-z]+://", definition.lower()):
         response = requests.get(definition, verify=(not should_disable_connection_verify()))
@@ -244,6 +250,7 @@ def _try_load_dashboard_definition(cmd, resource_group_name, grafana_name, defin
             definition = json.loads(response.content.decode())
         else:
             raise ArgumentUsageError(f"Failed to dashboard definition from '{definition}'. Error: '{response}'.")
+
     else:
         definition = json.loads(_try_load_file_content(definition))
 
@@ -255,9 +262,7 @@ def delete_dashboard(cmd, grafana_name, uid, resource_group_name=None):
 
 
 def create_data_source(cmd, grafana_name, definition, resource_group_name=None):
-    definition = _try_load_file_content(definition)
-    payload = json.loads(definition)
-    response = _send_request(cmd, resource_group_name, grafana_name, "post", "/api/datasources", payload)
+    response = _send_request(cmd, resource_group_name, grafana_name, "post", "/api/datasources", definition)
     return json.loads(response.content)
 
 
@@ -276,10 +281,47 @@ def list_data_sources(cmd, grafana_name, resource_group_name=None):
 
 
 def update_data_source(cmd, grafana_name, data_source, definition, resource_group_name=None):
-    definition = _try_load_file_content(definition)
     data = _find_data_source(cmd, resource_group_name, grafana_name, data_source)
     response = _send_request(cmd, resource_group_name, grafana_name, "put", "/api/datasources/" + str(data['id']),
-                             json.loads(definition))
+                             definition)
+    return json.loads(response.content)
+
+
+def list_notification_channels(cmd, grafana_name, resource_group_name=None, short=False):
+    if short is False:
+        response = _send_request(cmd, resource_group_name, grafana_name, "get", "/api/alert-notifications")
+    elif short is True:
+        response = _send_request(cmd, resource_group_name, grafana_name, "get", "/api/alert-notifications/lookup")
+    return json.loads(response.content)
+
+
+def show_notification_channel(cmd, grafana_name, notification_channel, resource_group_name=None):
+    return _find_notification_channel(cmd, resource_group_name, grafana_name, notification_channel)
+
+
+def create_notification_channel(cmd, grafana_name, definition, resource_group_name=None):
+    response = _send_request(cmd, resource_group_name, grafana_name, "post", "/api/alert-notifications", definition)
+    return json.loads(response.content)
+
+
+def update_notification_channel(cmd, grafana_name, notification_channel, definition, resource_group_name=None):
+    data = _find_notification_channel(cmd, resource_group_name, grafana_name, notification_channel)
+    definition['id'] = data['id']
+    response = _send_request(cmd, resource_group_name, grafana_name, "put",
+                             "/api/alert-notifications/" + str(data['id']),
+                             definition)
+    return json.loads(response.content)
+
+
+def delete_notification_channel(cmd, grafana_name, notification_channel, resource_group_name=None):
+    data = _find_notification_channel(cmd, resource_group_name, grafana_name, notification_channel)
+    _send_request(cmd, resource_group_name, grafana_name, "delete", "/api/alert-notifications/" + str(data["id"]))
+
+
+def test_notification_channel(cmd, grafana_name, notification_channel, resource_group_name=None):
+    data = _find_notification_channel(cmd, resource_group_name, grafana_name, notification_channel)
+    response = _send_request(cmd, resource_group_name, grafana_name, "post", "/api/alert-notifications/test",
+                             data)
     return json.loads(response.content)
 
 
@@ -412,6 +454,20 @@ def _find_data_source(cmd, resource_group_name, grafana_name, data_source):
                                      raise_for_error_status=False)
     if response.status_code >= 400:
         raise ArgumentUsageError(f"Couldn't found data source {data_source}. Ex: {response.status_code}")
+    return json.loads(response.content)
+
+
+def _find_notification_channel(cmd, resource_group_name, grafana_name, notification_channel):
+    response = _send_request(cmd, resource_group_name, grafana_name, "get",
+                             "/api/alert-notifications/" + notification_channel,
+                             raise_for_error_status=False)
+    if response.status_code >= 400:
+        response = _send_request(cmd, resource_group_name, grafana_name,
+                                 "get", "/api/alert-notifications/uid/" + notification_channel,
+                                 raise_for_error_status=False)
+    if response.status_code >= 400:
+        raise ArgumentUsageError(
+            f"Couldn't found notification channel {notification_channel}. Ex: {response.status_code}")
     return json.loads(response.content)