+++ src/containerapp/HISTORY.rst
@@ -3,9 +3,16 @@
 Release History
 ===============
 
-0.3.8
+0.3.9
 ++++++
 * 'az containerapp create': allow authenticating with managed identity (MSI) instead of ACR username & password
+
+0.3.8
+++++++
+* 'az containerapp update': Fix bug where --yaml would error out due to secret values
+* 'az containerapp update': use PATCH API instead of GET and PUT
+* 'az containerapp up': Fix bug where using --source with an invalid name parameter causes ACR build to fail
+* 'az containerapp logs show'/'az containerapp exec': Fix bug where ssh/logstream they would fail on apps with networking restrictions
 
 0.3.7
 ++++++